requirements.txt
# ABM Framework
mesa>=3.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
polars>=0.19.0
scipy>=1.11.0

# Performance Optimization
numba>=0.58.0

# Machine Learning & NLP
transformers>=4.35.0
torch>=2.0.0
vaderSentiment>=3.3.2
scikit-learn>=1.3.0

# API Clients
py-clob-client>=0.20.0
google-generativeai>=0.3.0
requests>=2.31.0

# Sports Data
# Note: nflverse data is accessed via direct parquet URLs, no package needed
sportradar>=1.0.0

# Kalshi API
kalshi-python>=1.0.0

# Backtesting
vectorbt>=0.26.0
quantstats>=0.0.62

# Visualization
solara>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Storage
influxdb-client>=1.38.0
redis>=5.0.0

# Async & WebSockets
websockets>=11.0
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Logging
structlog>=23.2.0

# Utilities
tenacity>=8.2.0
tqdm>=4.66.0
python-dateutil>=2.8.2

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
